refactor: refactor config relationship code for readability
